+++ Cargo.toml
@@ -11,7 +11,6 @@ homepage = "https://github.com/theatrus/psf-guard"
 readme = "README.md"
 keywords = ["astronomy", "telescope", "imaging", "analysis", "nina"]
 categories = ["command-line-utilities", "science"]
-
 [lib]
 name = "psf_guard"
 path = "src/lib.rs"
